main/requirements.txt
Flask>=1.1.2
Flask-SQLAlchemy>=2.4.4
SQLAlchemy>=1.3.20
Flask-Migrate>=2.5.3
Flask-Script>=2.0.6
Flask-Cors>=3.0.9
requests>=2.25.0
pika>=1.1.0
psycopg2-binary>=2.9
boto3
pymysql
cryptography
prometheus-flask-exporter
gunicorn
opentelemetry-api 
opentelemetry-sdk 
opentelemetry-instrumentation-flask 
opentelemetry-instrumentation-pika 
opentelemetry-exporter-jaeger-thrift